requirements.txt
crispy-bootstrap5==0.7
Django==4.1.7
django-braces==1.15.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==23.1
django-private-storage==3.0
django-pure-pagination==0.3.0
django-tinymce==3.6.1
docker==6.0.1
rules==3.3
timeago==1.0.16